Fixes #67 Moved the dialect query classes into their own module
